+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix: change xml namespace prefix `pagos10` to `pago10`
 - Refactor `CfdiUtils\Certificado\SerialNumber::baseConvert`
 - Add `CfdiUtils\Certificado\SerialNumber::asDecimal()`
+- Fix `CfdiUtils\Cleaner\Cleaner` since internal `DOMDocument` can be null
 
 
 # Version 2.4.4 2018-05-11